updating tests for droppable
