+++ test/unit/view/connected-droppable/disabled.spec.js
@@ -38,9 +38,6 @@ describe('home list', () => {
       impact: cloneImpact(noImpact),
     });
     const isHomeButNotOver: MapProps = {
-      isDraggingOver: false,
-      draggingOverWith: null,
-      draggingFromThisWith: preset.inHome1.descriptor.id,
       placeholder: preset.inHome1.placeholder,
       shouldAnimatePlaceholder: false,
       snapshot: {
