+++ test/unit/view/connected-droppable/dragging.spec.js
@@ -30,9 +30,6 @@ const state = getStatePreset();
 describe('home list', () => {
   const ownProps: OwnProps = getOwnProps(preset.home);
   const isOverMapProps: MapProps = {
-    isDraggingOver: true,
-    draggingOverWith: preset.inHome1.descriptor.id,
-    draggingFromThisWith: preset.inHome1.descriptor.id,
     placeholder: preset.inHome1.placeholder,
     shouldAnimatePlaceholder: false,
     snapshot: {
@@ -106,9 +103,6 @@ describe('home list', () => {
     });
 
     const isHomeButNotOver: MapProps = {
-      isDraggingOver: false,
-      draggingOverWith: null,
-      draggingFromThisWith: preset.inHome1.descriptor.id,
       placeholder: preset.inHome1.placeholder,
       shouldAnimatePlaceholder: false,
       snapshot: {
@@ -175,9 +169,6 @@ describe('foreign list', () => {
     };
 
     const isOverForeignMapProps: MapProps = {
-      isDraggingOver: true,
-      draggingFromThisWith: null,
-      draggingOverWith: preset.inHome1.descriptor.id,
       placeholder: preset.inHome1.placeholder,
       shouldAnimatePlaceholder: true,
       snapshot: {
