+++ test/unit/view/connected-droppable/util/resting-props.js
@@ -2,9 +2,6 @@
 import type { MapProps } from '../../../../../src/view/droppable/droppable-types';
 
 const restingProps: MapProps = {
-  isDraggingOver: false,
-  draggingOverWith: null,
-  draggingFromThisWith: null,
   placeholder: null,
   shouldAnimatePlaceholder: true,
   snapshot: {
